+++ packages/canonical-bridge-widget/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bnb-chain/canonical-bridge-widget
 
+## 0.7.0-alpha.1
+
+### Minor Changes
+
+- Fix price tag color
+
+## 0.7.0-alpha.0
+
+### Minor Changes
+
+- Add price change ratio tag
+
 ## 0.6.1
 
 ### Patch Changes